# prova/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Comandas",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/comandas": {
            "get": {
                "tags": [
                    "Comandas"
                ],
                "summary": "Listar usuários com suas comandas e produtos",
                "operationId": "99f1a009ba10ad39dbb8cbc0c3f1cf15",
                "responses": {
                    "200": {
                        "description": "Lista de usuários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "idUsuario": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nomeUsuario": {
                                                "type": "string",
                                                "example": "Maria"
                                            },
                                            "telefoneUsuario": {
                                                "type": "string",
                                                "example": "999999999"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comandas"
                ],
                "summary": "Criar uma nova comanda com usuário e produtos",
                "operationId": "6e2b0d567a4d7ccb2be748e5bcb599fe",
                "requestBody": {
                    "description": "Dados do usuário e produtos para a comanda",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idUsuario": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nomeUsuario": {
                                        "type": "string",
                                        "example": "Maria"
                                    },
                                    "telefoneUsuario": {
                                        "type": "string",
                                        "example": "999999999"
                                    },
                                    "produtos": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id",
                                                "nome",
                                                "preco"
                                            ],
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "nome": {
                                                    "type": "string",
                                                    "example": "Produto A"
                                                },
                                                "preco": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 9.99
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comanda criada com sucesso"
                    },
                    "400": {
                        "description": "Dados do produto incompletos"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/comandas/{id}": {
            "get": {
                "tags": [
                    "Comandas"
                ],
                "summary": "Detalhes do usuário com suas comandas e produtos",
                "operationId": "f7c129099280a3c018903a25b1c2063a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do usuário e seus produtos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "idUsuario": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nomeUsuario": {
                                            "type": "string",
                                            "example": "Maria"
                                        },
                                        "telefoneUsuario": {
                                            "type": "string",
                                            "example": "999999999"
                                        },
                                        "produtos": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "nome": {
                                                        "type": "string",
                                                        "example": "Produto A"
                                                    },
                                                    "preco": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 9.99
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Comandas"
                ],
                "summary": "Atualizar os produtos de uma comanda existente",
                "operationId": "d972a823d1c7b28a8b15507f19fb6730",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da comanda",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Lista de produtos para atualizar na comanda",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "produtos": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id",
                                                "nome",
                                                "preco"
                                            ],
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "nome": {
                                                    "type": "string",
                                                    "example": "Produto A"
                                                },
                                                "preco": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 9.99
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comanda atualizada com sucesso"
                    },
                    "404": {
                        "description": "Comanda não encontrada"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comandas"
                ],
                "summary": "Excluir uma comanda",
                "operationId": "3878520a9846b370b9a606417dd7c671",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da comanda",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comanda removida com sucesso"
                    },
                    "404": {
                        "description": "Comanda não encontrada"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/comandas/minha": {
            "get": {
                "tags": [
                    "Comandas"
                ],
                "summary": "Listar comandas e produtos do usuário autenticado",
                "operationId": "8c46482d14a93225d98eacac620687e3",
                "responses": {
                    "200": {
                        "description": "Lista de comandas com produtos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "comandas": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "produtos": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 10
                                                                },
                                                                "nome": {
                                                                    "type": "string",
                                                                    "example": "Produto A"
                                                                },
                                                                "preco": {
                                                                    "type": "number",
                                                                    "format": "float",
                                                                    "example": 9.99
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "API Token",
                "scheme": "bearer"
            }
        }
    },
...
}